Fix entry generation for classes implementing custom interfaces with registered members (#376)
* Fix #371 by checking if the overridden property is from an interface

Sub sequentially this also increases the amount of data the entry generator receives in order to do its job. Previously it only received registered classes, functions, properties and constructors. Now it receives every class, function and property from the symbol processor. This allows for additional checks on not registered members or not explicitly registered members.

* Implement check if overridden registered property is registered

* Make overridden registered member check a warning and not an error

* Remove implicit registration of overriding property if the overridden property is registered

* Fix compiler overridee check

* Update test classes for new enforced explicit registration