Fix api functions returning enums
Godot's api provides an id by which we map the enum. Previously we just assumed that id == ordinal. But this is not always the case (like in the `Key` enum).
This changes it to retrieve enums by id's using the `from` function we already generate in the enums and add it to the enums we wrote by hand.

This costs more performance as it's no longer a simple array access by index but rather a `first` approach with an filter by id function. But I see no other way for now other than allocating a static map for each enum which maps the id to the enum.