Remove generation and references to local resource path
This is done in preparation for freeform gdj files.
As gdj filenames are unique by design, the local path when resolving does thus not matter.
The initial gdj out dir is computed based on the users setting now when the files are generated.

This imposes one breaking change: Before, a library author could decide how the gdj files were generated (hierarchical or not) now the users setting is taken.