fix: allow godot-kotlin-jvm gradle plugin to participate in multi-module projects. (#549)
Previously, any attempt at applying the
godot-kotlin-jvm gradle plugin to a sub
project would fail with a NullPointerException
becuase "ideaExtension.project" returns
NULL for every gradle project except for
the root project. This update addresses
the issue by redirecting the task to the
root project. This also means that we
implicitly have to add the "idea" plugin
to the root project as well.

Co-authored-by: Martin Häusler <martin.haeusler@txture.io>