Update exporting.md
Document conflict with official export templates

If you attempt to export with both the official Godot export templates and the godot-kotlin-jvm export templates installed, Godot will default to using its own and the exported game will not be able to load any scene with Kotlin/Java code. This is avoided by setting the "Custom Template" fields in the export window.